apps/dbt-docs/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml
altair==5.5.0
    # via streamlit
annotated-types==0.7.0
    # via pydantic
asn1crypto==1.5.1
    # via snowflake-connector-python
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
blinker==1.9.0
    # via streamlit
boto3==1.38.34
    # via snowflake-connector-python
botocore==1.38.34
    # via
    #   boto3
    #   s3transfer
    #   snowflake-connector-python
cachetools==5.5.2
    # via streamlit
certifi==2025.4.26
    # via
    #   requests
    #   snowflake-connector-python
cffi==1.17.1
    # via
    #   cryptography
    #   snowflake-connector-python
charset-normalizer==3.4.2
    # via
    #   requests
    #   snowflake-connector-python
click==8.1.8
    # via
    #   snowflake-cli
    #   streamlit
    #   typer
cloudpickle==3.0.0
    # via snowflake-snowpark-python
cryptography==45.0.4
    # via
    #   pyopenssl
    #   snowflake-connector-python
filelock==3.18.0
    # via snowflake-connector-python
gitdb==4.0.12
    # via gitpython
gitpython==3.1.44
    # via
    #   snowflake-cli
    #   streamlit
idna==3.10
    # via
    #   requests
    #   snowflake-connector-python
jaraco-classes==3.4.0
    # via keyring
jaraco-context==6.0.1
    # via keyring
jaraco-functools==4.1.0
    # via keyring
jinja2==3.1.6
    # via
    #   altair
    #   pydeck
    #   snowflake-cli
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsonschema==4.24.0
    # via altair
jsonschema-specifications==2025.4.1
    # via jsonschema
keyring==25.6.0
    # via snowflake-connector-python
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
more-itertools==10.7.0
    # via
    #   jaraco-classes
    #   jaraco-functools
narwhals==1.42.0
    # via altair
numpy==2.3.0
    # via
    #   pandas
    #   pydeck
    #   streamlit
packaging==24.2
    # via
    #   altair
    #   requirements-parser
    #   snowflake-cli
    #   snowflake-connector-python
    #   streamlit
pandas==2.3.0
    # via streamlit
pillow==11.2.1
    # via streamlit
pip==25.1.1
    # via snowflake-cli
platformdirs==4.3.8
    # via snowflake-connector-python
pluggy==1.6.0
    # via snowflake-cli
prompt-toolkit==3.0.51
    # via snowflake-cli
protobuf==5.29.5
    # via
    #   snowflake-snowpark-python
    #   streamlit
pyarrow==18.1.0
    # via
    #   training-snowpark-container-service (pyproject.toml)
    #   streamlit
pycparser==2.22
    # via cffi
pydantic==2.11.4
    # via
    #   snowflake-cli
    #   snowflake-core
pydantic-core==2.33.2
    # via pydantic
pydeck==0.9.1
    # via streamlit
pygments==2.19.1
    # via rich
pyjwt==2.10.1
    # via snowflake-connector-python
pyopenssl==25.1.0
    # via snowflake-connector-python
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   pandas
    #   snowflake-core
    #   snowflake-snowpark-python
pytz==2025.2
    # via
    #   pandas
    #   snowflake-connector-python
pyyaml==6.0.2
    # via
    #   snowflake-cli
    #   snowflake-core
    #   snowflake-snowpark-python
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via
    #   snowflake-cli
    #   snowflake-connector-python
    #   snowflake-core
    #   streamlit
requirements-parser==0.11.0
    # via snowflake-cli
rich==14.0.0
    # via
    #   snowflake-cli
    #   typer
rpds-py==0.25.1
    # via
    #   jsonschema
    #   referencing
s3transfer==0.13.0
    # via boto3
setuptools==80.3.1
    # via
    #   snowflake-cli
    #   snowflake-snowpark-python
shellingham==1.5.4
    # via typer
six==1.17.0
    # via python-dateutil
smmap==5.0.2
    # via gitdb
snowflake-cli==3.9.1
    # via snowflake-cli-labs
snowflake-cli-labs==3.9.1
    # via training-snowpark-container-service (pyproject.toml)
snowflake-connector-python==3.15.0
    # via
    #   training-snowpark-container-service (pyproject.toml)
    #   snowflake-cli
    #   snowflake-core
    #   snowflake-snowpark-python
snowflake-core==1.4.0
    # via snowflake-cli
snowflake-snowpark-python==1.32.0
    # via training-snowpark-container-service (pyproject.toml)
sortedcontainers==2.4.0
    # via snowflake-connector-python
streamlit==1.45.1
    # via training-snowpark-container-service (pyproject.toml)
tenacity==9.1.2
    # via streamlit
toml==0.10.2
    # via streamlit
tomlkit==0.13.2
    # via
    #   snowflake-cli
    #   snowflake-connector-python
tornado==6.5.1
    # via streamlit
typer==0.15.2
    # via snowflake-cli
types-setuptools==80.9.0.20250529
    # via requirements-parser
typing-extensions==4.14.0
    # via
    #   altair
    #   pydantic
    #   pydantic-core
    #   snowflake-connector-python
    #   snowflake-snowpark-python
    #   streamlit
    #   typer
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
tzdata==2025.2
    # via pandas
tzlocal==5.3.1
    # via snowflake-snowpark-python
urllib3==2.4.0
    # via
    #   botocore
    #   requests
    #   snowflake-cli
    #   snowflake-core
wcwidth==0.2.13
    # via prompt-toolkit
wheel==0.45.1
    # via snowflake-snowpark-python
